Create README.md
Cinema Ticket booking system: Create a Java Application to book a ticket in a Cinema Hall. Create appropriate class to represent a Ticket and User. Use 2-D arrays to store the seats of the hall.
Definition of done:
xiv. The System should ask user to enter Username and Password.
XV. The system should give option of different available movies. After selecting a movie give option of different seats available.
xvi. A user should be able to book or cancel a ticket.
xvii. Store the information in text file.